refactor(cli-utils): refactor cli utils structure
