+++ packages/utils/cli/src/core/Logger.mjs
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 import chalk from 'chalk'
 
 export class Logger {
