+++ packages/utils/cli/src/generate/templates/component/[.npmignore].js
