+++ packages/utils/cli/src/generate/templates/component/[package.json].js
