+++ packages/utils/cli/src/generate/templates/component/[tsconfig.json].js
