+++ packages/utils/cli/src/generate/templates/component/[vite.config.ts].js
