+++ packages/utils/cli/src/generate/templates/component/src/[Component.stories.mdx].js
