+++ packages/utils/cli/src/generate/templates/component/src/[Component.test.tsx].js
