+++ packages/utils/cli/src/generate/templates/component/src/[Component.tsx].js
