+++ packages/utils/cli/src/generate/templates/component/src/[Component.variants.tsx].js
