+++ packages/utils/cli/src/generate/templates/component/src/[index.ts].js
