+++ packages/utils/cli/src/generate/templates/hook/[.npmignore].js
