+++ packages/utils/cli/src/generate/templates/hook/[package.json].js
