+++ packages/utils/cli/src/generate/templates/hook/[tsconfig.json].js
