+++ packages/utils/cli/src/generate/templates/hook/[vite.config.ts].js
