+++ packages/utils/cli/src/generate/templates/hook/src/[index.ts].js
