+++ packages/utils/cli/src/generate/templates/hook/src/[name.stories.mdx].js
