+++ packages/utils/cli/src/generate/templates/hook/src/[name.test.tsx].js
