+++ packages/utils/cli/src/generate/templates/hook/src/[name.tsx].js
