+++ packages/utils/cli/src/generate/templates/util/[.npmignore].js
