+++ packages/utils/cli/src/generate/templates/util/[package.json].js
