+++ packages/utils/cli/src/generate/templates/util/[tsconfig.json].js
