+++ packages/utils/cli/src/generate/templates/util/[vite.config.ts].js
