+++ packages/utils/cli/src/generate/templates/util/src/[index.ts].js
