+++ packages/utils/cli/src/generate/validators/index.mjs
