+++ packages/utils/cli/src/setup-theme/utils/index.js
